+++ src/main/java/insane96mcp/progressivebosses/module/wither/WitherFeature.java
@@ -27,9 +27,10 @@ import java.util.List;
 @Label(name = "Wither Feature")
 @LoadFeature(module = ProgressiveBosses.RESOURCE_PREFIX + "wither", canBeDisabled = false)
 public class WitherFeature extends Feature {
-    public static final TagKey<Item> WITHER_INVULNERABLE = ItemTags.create(new ResourceLocation(ProgressiveBosses.MOD_ID, "wither_invulnerable"));
+	public static final TagKey<Item> WITHER_INVULNERABLE = ItemTags.create(new ResourceLocation(ProgressiveBosses.MOD_ID, "wither_invulnerable"));
+	public static final TagKey<Item> WORLD_INVULNERABLE = ItemTags.create(new ResourceLocation(ProgressiveBosses.MOD_ID, "world_invulnerable"));
 
-    public WitherFeature(Module module, boolean enabledByDefault, boolean canBeDisabled) {
+	public WitherFeature(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
 	}